ContextEntryCell uses ContextEntry for rendering its content
