+++ kie-wb-common-react/boxed-expression-editor/src/components/ContextExpression/ContextEntryCell.css
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,22 +14,15 @@
  * limitations under the License.
  */
 
-.expression-container .context-expression .context-result {
-  display: flex;
-}
-
-.expression-container .context-expression .context-result .result-label {
-  display: flex;
-  flex: 0.5 1;
-  border-right-style: ridge;
-  padding: 0.5em 0.5em 0 0.5em;
+.expression-container .context-expression .context-entry .entry-definition {
   cursor: pointer;
-  font-style: italic;
-  justify-content: center;
-  flex-direction: column;
-  color: gray;
 }
 
-.expression-container .context-expression .context-result .result-expression {
-  flex: 2 1;
+.expression-container .context-expression .context-entry .entry-definition .entry-name {
+  white-space: nowrap;
+}
+
+.expression-container .context-expression .context-entry .entry-definition .entry-data-type {
+  font-style: italic;
+  font-size: smaller;
 }
