+++ kie-wb-common-react/boxed-expression-editor/src/components/ContextExpression/ContextExpression.css
@@ -14,6 +14,11 @@
  * limitations under the License.
  */
 
-export * from "./ContextExpression";
-export * from "./ContextEntryCell";
-export * from "./ContextEntry";
+.context-expression {
+  width: 100%;
+}
+
+.context-expression .context-result {
+  font-style: italic;
+  color: gray;
+}
